Gemini: Attempted implementation of protocol
Incomplete and currently non-functional implementation of Gemini
protocol. Uses QSslSocket now, and correctly forms the request
message. Does not currently handle the response correctly.
Problem: it doesn't seem to be raising the sslErrors or
peerVerifyError signals, it just skips straight to disconnecting,
and will list SSLHandshakeError -9824 if the errorString is printed
during the disconnect state. This is the errSSLPeerHandshakeFail
error and seems to be coming directly from Mac OS