Gemini: Basic functional impl. of protocol+spec
This commit represents hours of annoyance. The SSL issues mentioned in
my previous commits, and the evidence of which is scattered all over the
code now, are still not solved. I am simply refusing to acknowledge the
problem now, I'm just built different I guess.
In reality, I can't do anything about the SSL issues right now because
it transpires that the root issue is outside of Qt's QSslSocket class.
I can't work out how to indicate that the SSL certificate is not
necessarily going to be signed by a recognised CA, but that if it has
not been encoutered before, it should be trusted until changed, a-la
OpenSSH (as discussed in the Gemini spec, this is referred to as TOFU).
In the interim, I've found, and been testing spec compliance against, a
lovely person's website, gemini://mozz.us

I'm working with some very broken recollection of C++ and literal
guesswork in terms of Qt, so please don't judge too harshly.